+++ legacy/callbacks/mod.rs
@@ -1,4 +1,3 @@
-pub mod designer;
 pub mod editor;
 pub mod settings;
 pub mod machine;