+++ src/app/callbacks/mod.rs
@@ -1,5 +1,0 @@
-pub mod designer;
-pub mod editor;
-pub mod settings;
-pub mod machine;
-pub mod cam;